+++ ReplaceSpacesByPercent20/ReplaceSpacesByPercent20/ReplaceSpacesByPercent20.cpp
@@ -10,7 +10,7 @@ int _tmain(int argc, _TCHAR* argv[])
 	char statement[100] = " Should replace  spaces by percentage20 ";
 	int numberOfSpaces = 0, newSize;
 
-	for (int i = 0; i < strlen(statement); ++i)
+	for (size_t i = 0; i < strlen(statement); ++i)
 	{
 		if (statement[i] == ' ')
 		{
